requirements.txt
azure-ai-formrecognizer==3.3.0
pandas==2.1.4
Pillow==10.1.0
python-dotenv==1.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
spacy==3.7.2
PyPDF2==3.0.1
requests==2.31.0
streamlit==1.28.2
reportlab==4.0.4 